+++ lib/metrics/http.go
@@ -25,9 +25,6 @@ import (
 	"fmt"
 	"net/http"
 	"runtime"
-	"strings"
-	"sync"
-	"sync/atomic"
 	"time"
 
 	"github.com/Jeffail/benthos/lib/log"
@@ -54,61 +51,23 @@ var (
 
 //------------------------------------------------------------------------------
 
-// HTTPStat is a representation of a single metric stat. Interactions with this
-// stat are thread safe.
-type HTTPStat struct {
-	value *int64
-}
-
-// Incr increments a metric by an amount.
-func (h *HTTPStat) Incr(count int64) error {
-	atomic.AddInt64(h.value, count)
-	return nil
-}
-
-// Decr decrements a metric by an amount.
-func (h *HTTPStat) Decr(count int64) error {
-	atomic.AddInt64(h.value, -count)
-	return nil
-}
-
-// Timing sets a timing metric.
-func (h *HTTPStat) Timing(delta int64) error {
-	atomic.StoreInt64(h.value, delta)
-	return nil
-}
-
-// Gauge sets a gauge metric.
-func (h *HTTPStat) Gauge(value int64) error {
-	atomic.StoreInt64(h.value, value)
-	return nil
-}
-
-//------------------------------------------------------------------------------
-
 // HTTP is an object with capability to hold internal stats as a JSON endpoint.
 type HTTP struct {
-	flatCounters map[string]*int64
-	flatTimings  map[string]*int64
-	pathPrefix   string
-	timestamp    time.Time
-
-	sync.Mutex
+	local      *Local
+	timestamp  time.Time
+	pathPrefix string
 }
 
 // NewHTTP creates and returns a new HTTP object.
 func NewHTTP(config Config, opts ...func(Type)) (Type, error) {
 	t := &HTTP{
-		flatCounters: map[string]*int64{},
-		flatTimings:  map[string]*int64{},
-		pathPrefix:   config.Prefix,
-		timestamp:    time.Now(),
+		local:      NewLocal(),
+		timestamp:  time.Now(),
+		pathPrefix: config.Prefix,
 	}
-
 	for _, opt := range opts {
 		opt(t)
 	}
-
 	return t, nil
 }
 
@@ -120,23 +79,14 @@ func (h *HTTP) HandlerFunc() http.HandlerFunc {
 		uptime := time.Since(h.timestamp).String()
 		goroutines := runtime.NumGoroutine()
 
-		h.Lock()
-		localFlatCounters := make(map[string]int64, len(h.flatCounters))
-		for k := range h.flatCounters {
-			localFlatCounters[k] = atomic.LoadInt64(h.flatCounters[k])
-		}
-
-		localFlatTimings := make(map[string]int64, len(h.flatTimings))
-		for k := range h.flatTimings {
-			localFlatTimings[k] = atomic.LoadInt64(h.flatTimings[k])
-		}
-		h.Unlock()
+		counters := h.local.GetCounters()
+		timings := h.local.GetTimings()
 
 		obj := gabs.New()
-		for k, v := range localFlatCounters {
+		for k, v := range counters {
 			obj.SetP(v, k)
 		}
-		for k, v := range localFlatTimings {
+		for k, v := range timings {
 			obj.SetP(v, k)
 			obj.SetP(time.Duration(v).String(), k+"_readable")
 		}
@@ -156,108 +106,37 @@ func (h *HTTP) HandlerFunc() http.HandlerFunc {
 
 // GetCounter returns a stat counter object for a path.
 func (h *HTTP) GetCounter(path ...string) StatCounter {
-	dotPath := strings.Join(path, ".")
-
-	h.Lock()
-	ptr, exists := h.flatCounters[dotPath]
-	if !exists {
-		var ctr int64
-		ptr = &ctr
-		h.flatCounters[dotPath] = ptr
-	}
-	h.Unlock()
-
-	return &HTTPStat{
-		value: ptr,
-	}
+	return h.local.GetCounter(path...)
 }
 
 // GetTimer returns a stat timer object for a path.
 func (h *HTTP) GetTimer(path ...string) StatTimer {
-	dotPath := strings.Join(path, ".")
-
-	h.Lock()
-	ptr, exists := h.flatTimings[dotPath]
-	if !exists {
-		var ctr int64
-		ptr = &ctr
-		h.flatTimings[dotPath] = ptr
-	}
-	h.Unlock()
-
-	return &HTTPStat{
-		value: ptr,
-	}
+	return h.local.GetTimer(path...)
 }
 
 // GetGauge returns a stat gauge object for a path.
 func (h *HTTP) GetGauge(path ...string) StatGauge {
-	dotPath := strings.Join(path, ".")
-
-	h.Lock()
-	ptr, exists := h.flatCounters[dotPath]
-	if !exists {
-		var ctr int64
-		ptr = &ctr
-		h.flatCounters[dotPath] = ptr
-	}
-	h.Unlock()
-
-	return &HTTPStat{
-		value: ptr,
-	}
+	return h.local.GetGauge(path...)
 }
 
 // Incr increments a stat by a value.
 func (h *HTTP) Incr(stat string, value int64) error {
-	h.Lock()
-	if ptr, exists := h.flatCounters[stat]; !exists {
-		ctr := value
-		h.flatCounters[stat] = &ctr
-	} else {
-		atomic.AddInt64(ptr, value)
-	}
-	h.Unlock()
-	return nil
+	return h.local.Incr(stat, value)
 }
 
 // Decr decrements a stat by a value.
 func (h *HTTP) Decr(stat string, value int64) error {
-	h.Lock()
-	if ptr, exists := h.flatCounters[stat]; !exists {
-		ctr := -value
-		h.flatCounters[stat] = &ctr
-	} else {
-		atomic.AddInt64(ptr, -value)
-	}
-	h.Unlock()
-	return nil
+	return h.local.Decr(stat, value)
 }
 
 // Timing sets a stat representing a duration.
 func (h *HTTP) Timing(stat string, delta int64) error {
-	h.Lock()
-	if ptr, exists := h.flatTimings[stat]; !exists {
-		ctr := delta
-		h.flatTimings[stat] = &ctr
-	} else {
-		atomic.StoreInt64(ptr, delta)
-	}
-	h.Unlock()
-	return nil
+	return h.local.Timing(stat, delta)
 }
 
 // Gauge sets a stat as a gauge value.
 func (h *HTTP) Gauge(stat string, value int64) error {
-	h.Lock()
-	if ptr, exists := h.flatCounters[stat]; !exists {
-		ctr := value
-		h.flatTimings[stat] = &ctr
-	} else {
-		atomic.StoreInt64(ptr, value)
-	}
-	h.Unlock()
-	return nil
+	return h.local.Gauge(stat, value)
 }
 
 // SetLogger does nothing.